requirements.txt
streamlit
langchain-core
langchain-google-genai
langgraph
mcp
sqlmodel
python-dotenv
pydantic
zscaler
langchain-mcp-adapters
pycountry
uvicorn
Pillow
